+++ example/utp_close_example.dart
@@ -1,8 +1,6 @@
-import 'dart:async';
 import 'dart:io';
 import 'dart:typed_data';
 
-import 'package:test/test.dart';
 import 'package:utp/src/utp_protocol_implement.dart';
 
 void main(List<String> args) async {